requirements.txt
wheel
mcipc
discord.py
python-dotenv
psycopg2
requests